More pretty responses

+++ app/controllers/application_controller.rb
@@ -25,7 +25,7 @@ class ApplicationController < ActionController::Base
           from_number = PhoneNumber.find_by_number!(from)
           to_number = PhoneNumber.find_by_number!(to)
 
-          raise "Same Person" if from_number.number == to_number.number
+          raise "Same Person. Send 'help' for assistance." if from_number.number == to_number.number
 
           from_wallet = WalletThang.new(from_number)
           to_wallet = WalletThang.new(to_number)
@@ -49,7 +49,7 @@ class ApplicationController < ActionController::Base
           from_number = PhoneNumber.find_by_number!(from)
           to_number = PhoneNumber.find_by_number!(to)
 
-          raise "same person" if from_number.number == to_number.number
+          raise "Same Person. Send 'help' for assistance." if from_number.number == to_number.number
 
           #from_wallet = WalletThang.new(from_number)
           #to_wallet = WalletThang.new(to_number)
@@ -64,13 +64,13 @@ class ApplicationController < ActionController::Base
           $twilio_client.account.sms.messages.create(:from => mega_from, :to => from_number.number, :body => "You are requesting #{amount} from #{to_number.number} for #{thing}.")
           $twilio_client.account.sms.messages.create(:from => mega_from, :to => to_number.number, :body => "#{from_number.number} is requesting #{amount} for #{thing}. Confirm or Deny?")
 
-        when "ok"
+        when "ok", "yes", "confirm"
           raise "invalid syntax" unless parts.length == 1
 
           from_number = PhoneNumber.find_by_number!(from)
           latest_request = Request.where(:to_id => from_number.id).order("created_at DESC").limit(1).first
 
-          raise "unknown request" unless latest_request
+          raise "Unknown Request. Send 'help' for assistance. " unless latest_request
 
           $twilio_client.account.sms.messages.create(
             :from => mega_from,
@@ -82,6 +82,28 @@ class ApplicationController < ActionController::Base
             :from => mega_from,
             :to => latest_request.from_phone_number.number,
             :body => "#{latest_request.to_phone_number.number} has accepted your #{latest_request.amount} request for #{latest_request.thing}"
+          )
+
+          Request.where(:to_id => from_number.id).destroy_all
+
+        when "no", "deny"
+          raise "invalid syntax" unless parts.length == 1
+
+          from_number = PhoneNumber.find_by_number!(from)
+          latest_request = Request.where(:to_id => from_number.id).order("created_at DESC").limit(1).first
+
+          raise "Unknown Request. Send 'help' for assistance. " unless latest_request
+
+          $twilio_client.account.sms.messages.create(
+              :from => mega_from,
+              :to => latest_request.to_phone_number.number,
+              :body => "You have denied sending #{latest_request.amount} to #{latest_request.from_phone_number.number} for #{latest_request.thing}"
+          )
+
+          $twilio_client.account.sms.messages.create(
+              :from => mega_from,
+              :to => latest_request.from_phone_number.number,
+              :body => "#{latest_request.to_phone_number.number} has denied your request of #{latest_request.amount} for #{latest_request.thing}"
           )
 
           Request.where(:to_id => from_number.id).destroy_all
